rename server controllers
