+++ Server/src/controller/Common/countVote.js
